# components/tenant-api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handler.BatchOperation:
    properties:
      patch:
        additionalProperties:
          type: string
        description: 'key → value to set (e.g., "_silent_mode": "warning")'
        type: object
      tenant_id:
        maxLength: 256
        minLength: 1
        type: string
    required:
    - tenant_id
    type: object
  internal_handler.BatchRequest:
    properties:
      operations:
        items:
          $ref: '#/definitions/internal_handler.BatchOperation'
        maxItems: 1000
        minItems: 1
        type: array
    required:
    - operations
    type: object
  internal_handler.BatchResponse:
    properties:
      message:
        description: 'v2.6.0: human-readable message'
        type: string
      pr_number:
        description: 'v2.6.0: PR/MR number in PR mode'
        type: integer
      pr_url:
        description: 'v2.6.0: PR/MR URL in PR mode'
        type: string
      results:
        items:
          $ref: '#/definitions/internal_handler.BatchResult'
        type: array
      status:
        description: '"completed" | "pending_review" (PR mode)'
        type: string
      summary:
        description: e.g., "5 succeeded, 1 failed"
        type: string
      task_id:
        description: async task ID
        type: string
    type: object
  internal_handler.BatchResult:
    properties:
      message:
        type: string
      status:
        description: '"ok" | "error"'
        type: string
      tenant_id:
        type: string
    type: object
...
  internal_handler.GroupBatchRequest:
    properties:
      patch:
        additionalProperties:
          type: string
        description: 'key → value (e.g., "_silent_mode": "warning")'
        type: object
    type: object
  internal_handler.GroupBatchResponse:
    properties:
      group_id:
        type: string
      results:
        items:
          $ref: '#/definitions/internal_handler.BatchResult'
        type: array
      status:
        type: string
      summary:
        description: e.g., "5 succeeded, 1 failed"
        type: string
      task_id:
        type: string
    type: object
  internal_handler.GroupResponse:
    properties:
      description:
        type: string
      filters:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      label:
        type: string
      members:
        items:
          type: string
        type: array
    type: object
...
  internal_handler.PutGroupRequest:
    properties:
      description:
        maxLength: 4096
        type: string
      filters:
        additionalProperties:
          type: string
        type: object
      label:
        maxLength: 256
        minLength: 1
        type: string
      members:
        items:
          type: string
        maxItems: 1000
        type: array
    required:
    - label
    type: object
...
info:
  contact: {}
  description: |-
    Multi-tenant configuration management API for the Dynamic Alerting Platform.
    Provides CRUD over per-tenant alert thresholds, RBAC-scoped writes, and async batch operations.
    Auth is enforced by the oauth2-proxy sidecar and forwarded via X-Forwarded-Email / X-Forwarded-Groups.
  license:
    name: See repository LICENSE
  title: Tenant API
  version: v2.8.0
paths:
...
  /api/v1/groups/{id}:
    delete:
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a group
      tags:
      - groups
    get:
      description: Returns a tenant group by ID.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.GroupResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a single group
      tags:
      - groups
    put:
      consumes:
      - application/json
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Group definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.PutGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create or update a group
      tags:
      - groups
  /api/v1/groups/{id}/batch:
    post:
      consumes:
      - application/json
      description: Apply a patch to all tenants in a group.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Patch to apply
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.GroupBatchRequest'
      - description: Enable async mode (true/false)
        in: query
        name: async
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.GroupBatchResponse'
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Batch operation on group members
      tags:
      - groups
...
  /api/v1/tenants/batch:
    post:
      consumes:
      - application/json
      description: Apply patch operations to multiple tenants in one call.
      parameters:
      - description: Batch operations
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.BatchRequest'
      - description: Enable async mode (true/false)
        in: query
        name: async
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.BatchResponse'
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Batch tenant operations
      tags:
      - tenants
...
swagger: "2.0"